+++ CHANGELOG.rst
@@ -25,11 +25,16 @@ Bug fixes
 * ``xs.io.save_to_table`` now correctly handles the case where the input is a `DataArray` or a `Dataset` with a single variable. (:pull:`473`).
 * Fixed a bug in ``xs.utils.change_units`` where the original dataset was also getting modified. (:pull:`482`).
 * Fixed a bug in ``xs.compute_indicators`` where the `cat:variable` attribute was not correctly set. (:pull:`483`).
+* Fixed a bug in ``xs.climatological_op`` where kwargs were not passed to the operation function. (:pull:`486`).
+* Fixed a bug in ``xs.climatological_op`` where `min_periods` was not passed when the operation was `linregress`. (:pull:`486`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^
 * Include CF convention for temperature differences and on scale (:pull:`428`, :issue:`428`).
 * Bumped the version of `xclim` to 0.53.2. (:pull:`482`).
+* More tests added. (:pull:`486`).
+* Fixed a bug in ``xs.testing.datablock_3d`` where some attributes of the rotated pole got reversed half-way through the creation of the dataset. (:pull:`486`).
+* The function ``xs.regrid._get_grid_mapping`` was moved to ``xs.spatial.get_grid_mapping`` and is now a public function. (:pull:`486`).
 
 v0.10.0 (2024-09-30)
 --------------------